:truck: testUtils move to test
